requirements.txt
streamlit
langchain-nvidia-ai-endpoints
python-dotenv
langchain-community
langchain
ipykernel